Updated to float32 for UserInfo properties.
Balance & Rating

+++ fivesim/responses.go
@@ -16,8 +16,8 @@ type Products map[string]Product
 type UserInfo struct {
 	ID      int     `json:"id"`
 	Email   string  `json:"email"`
-	Balance float64 `json:"balance"`
-	Rating  float64  `json:"rating"`
+	Balance float32 `json:"balance"`
+	Rating  float32  `json:"rating"`
 }
 
 // SMS represents info about an incoming SMS
